Lesson 4: PermCheck
Check whether array A is a permutation.

This was the last version, in this I considered other kind of solution.
I decided to have the correct sequence [1, 2, 3, 4] and subtract the values of the
given array [1, 3, 2, 4] and add these differences, the answer of that would be zero
if the given array is a permutation, but I added a function for avoiding repeated values too
and other to check was is the max value of permutation for example [1, 2, 3, 4]
the max value must to be 4, because the length of the array was 4.
This method gets 100%