Lesson 4: MissingInteger
Find the smallest positive integer that does not occur in a given sequence.

I got 100%! firstable I was thinking and how to reduce the iterations of
the for loop. I tried to compare two values at the same iteration but it
was confused then I remembered that it's better two separate loops than
one loop into another loop and using a set I would avoid duplicate values
reducing the iterations.